// Participium/server/src/models/dto/Report.ts
export interface Report {
    /**
     * 
     * @type {number}
     * @memberof Report
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Report
     */
    title?: string;
    /**
     * 
     * @type {Location}
     * @memberof Report
     */
    location?: Location;
    /**
     * 
     * @type {User}
     * @memberof Report
     */
    author?: User;
    /**
     * 
     * @type {boolean}
     * @memberof Report
     */
    anonymity?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Report
     */
    date?: string;
    /**
     * 
     * @type {OfficeType}
     * @memberof Report
     */
    category?: OfficeType;
    /**
     * 
     * @type {ReportDocument}
     * @memberof Report
     */
    document?: ReportDocument;
    /**
     * Current state of the report
     * @type {string}
     * @memberof Report
     */
    state?: string;
    /**
     * Id of the officer assigned to the report (if any)
     * @type {number}
     * @memberof Report
     */
    assignedOfficerId?: number;
    /**
     * Id of the maintainer assigned to the report (if any)
     * @type {number}
     * @memberof Report
     */
    assignedMaintainerId?: number;
    /**
     * Optional reason (for declines etc.)
     * @type {string}
     * @memberof Report
     */
    reason?: string;
    /**
     * Users that follow this report
     * @type {Array<User>}
     * @memberof Report
     */
    followerUsers?: Array<User>;
}